perception: moving lazy_free_space_updater into it's own library
