+++ perception/depth_image_octomap_updater/CMakeLists.txt
@@ -1,13 +1,10 @@
 set(MOVEIT_LIB_NAME moveit_depth_image_octomap_updater)
 
 add_library(${MOVEIT_LIB_NAME}_core
-  src/lazy_free_space_updater.cpp
   src/depth_image_octomap_updater.cpp
   )
 
-set_target_properties(${MOVEIT_LIB_NAME}_core PROPERTIES COMPILE_FLAGS "${CMAKE_CXX_FLAGS} ${OpenMP_CXX_FLAGS}")
-set_target_properties(${MOVEIT_LIB_NAME}_core PROPERTIES LINK_FLAGS "${OpenMP_CXX_FLAGS}")
-target_link_libraries(${MOVEIT_LIB_NAME}_core moveit_mesh_filter moveit_occupancy_map_monitor ${catkin_LIBRARIES} ${Boost_LIBRARIES})
+target_link_libraries(${MOVEIT_LIB_NAME}_core moveit_lazy_free_space_updater moveit_mesh_filter moveit_occupancy_map_monitor ${catkin_LIBRARIES} ${Boost_LIBRARIES})
 
 add_dependencies(${MOVEIT_LIB_NAME}_core sensor_msgs_generate_cpp)
