+++ perception/depth_image_octomap_updater/include/moveit/depth_image_octomap_updater/depth_image_octomap_updater.h
@@ -42,7 +42,7 @@
 #include <moveit/occupancy_map_monitor/occupancy_map_updater.h>
 #include <moveit/mesh_filter/mesh_filter.h>
 #include <moveit/mesh_filter/stereo_camera_model.h>
-#include <moveit/depth_image_octomap_updater/lazy_free_space_updater.h>
+#include <moveit/lazy_free_space_updater/lazy_free_space_updater.h>
 #include <image_transport/image_transport.h>
 #include <boost/scoped_ptr.hpp>
